use gl-matrix for vector manipulation; rename components for clarity
